+++ reservoir-benchmark/src/main/java/com/github/ylgrgyq/reservoir/benchmark/BenchmarkRunner.java
@@ -1,4 +1,4 @@
-package com.github.ylgrgyq.reservoir.benchmark.storage;
+package com.github.ylgrgyq.reservoir.benchmark;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
