+++ reservoir-benchmark/src/main/java/com/github/ylgrgyq/reservoir/benchmark/EnvironmentInfo.java
@@ -1,4 +1,4 @@
-package com.github.ylgrgyq.reservoir.benchmark.storage;
+package com.github.ylgrgyq.reservoir.benchmark;
 
 import java.lang.management.GarbageCollectorMXBean;
 import java.lang.management.ManagementFactory;
